requirements.txt
python-dateutil==2.8.2
hashids==1.3.1
requests~=2.21.0
python-decouple==3.3